backend/requirements.txt
fastapi
uvicorn
python-multipart
dnspython
dkimpy
requests
beautifulsoup4
lxml
